+++ src/entity-manager/EntityManager.ts
@@ -35,6 +35,7 @@ import { IsolationLevel } from "../driver/types/IsolationLevel"
 import { ObjectUtils } from "../util/ObjectUtils"
 import { getMetadataArgsStorage } from "../globals"
 import { UpsertOptions } from "../repository/UpsertOptions"
+import { UpdateOptions } from "../repository/UpdateOptions"
 import { InstanceChecker } from "../util/InstanceChecker"
 import { ObjectLiteral } from "../common/ObjectLiteral"
 import { PickKeysByType } from "../common/PickKeysByType"
@@ -740,7 +741,7 @@ export class EntityManager {
                 ),
         )
 
-        return this.createQueryBuilder()
+        const qb = this.createQueryBuilder()
             .insert()
             .into(target)
             .values(entities)
@@ -758,7 +759,12 @@ export class EntityManager {
                         this.connection.driver.supportedUpsertTypes[0],
                 },
             )
-            .execute()
+
+        if (options.returning !== undefined) {
+            qb.returning(options.returning)
+        }
+
+        return qb.execute()
     }
 
     /**
@@ -781,6 +787,7 @@ export class EntityManager {
             | ObjectId[]
             | any,
         partialEntity: QueryDeepPartialEntity<Entity>,
+        options?: UpdateOptions,
     ): Promise<UpdateResult> {
         // if user passed empty criteria or empty list of criterias, then throw an error
         if (OrmUtils.isCriteriaNullOrEmpty(criteria)) {
@@ -792,17 +799,27 @@ export class EntityManager {
         }
 
         if (OrmUtils.isPrimitiveCriteria(criteria)) {
-            return this.createQueryBuilder()
+            const qb = this.createQueryBuilder()
                 .update(target)
                 .set(partialEntity)
                 .whereInIds(criteria)
-                .execute()
+
+            if (options?.returning !== undefined) {
+                qb.returning(options.returning)
+            }
+
+            return qb.execute()
         } else {
-            return this.createQueryBuilder()
+            const qb = this.createQueryBuilder()
                 .update(target)
                 .set(partialEntity)
                 .where(criteria)
-                .execute()
+
+            if (options?.returning !== undefined) {
+                qb.returning(options.returning)
+            }
+
+            return qb.execute()
         }
     }
 
@@ -816,11 +833,15 @@ export class EntityManager {
     updateAll<Entity extends ObjectLiteral>(
         target: EntityTarget<Entity>,
         partialEntity: QueryDeepPartialEntity<Entity>,
+        options?: UpdateOptions,
     ): Promise<UpdateResult> {
-        return this.createQueryBuilder()
-            .update(target)
-            .set(partialEntity)
-            .execute()
+        const qb = this.createQueryBuilder().update(target).set(partialEntity)
+
+        if (options?.returning !== undefined) {
+            qb.returning(options.returning)
+        }
+
+        return qb.execute()
     }
 
     /**
@@ -1357,7 +1378,7 @@ export class EntityManager {
         const values: QueryDeepPartialEntity<Entity> = propertyPath
             .split(".")
             .reduceRight(
-                (value, key) => ({ [key]: value } as any),
+                (value, key) => ({ [key]: value }) as any,
                 () =>
                     this.connection.driver.escape(column.databaseName) +
                     " + " +
@@ -1394,7 +1415,7 @@ export class EntityManager {
         const values: QueryDeepPartialEntity<Entity> = propertyPath
             .split(".")
             .reduceRight(
-                (value, key) => ({ [key]: value } as any),
+                (value, key) => ({ [key]: value }) as any,
                 () =>
                     this.connection.driver.escape(column.databaseName) +
                     " - " +